tools: substitute ubsan_minimal for fuzzer_test in create-openbsd
Reliance on -fsanitize-coverage=trace-pc makes clang look for full
ubsan libraries. The only one currently available on OpenBSD is
ubsan_minimal. We temporarily use it in place of ubsan_standalone.
The ubsan_standalone_cxx library seems currently not necessary
and so an empty file works in its stead.

This is a temporary measure until I figure out the best course of
action for making fsanitize-coverage=trace-pc work out of the box
on OpenBSD.

This should help with one of the issues we just discovered having
replaced the base OS image we use to run syzkaller on OpenBSD.