vm: fix logical race in a test
Currently the inject-error test injects "BUG: foo\n"
and expects VM to fail with this error.
However, the command immidiatly exists, so it may exist
before the output is injected. In that case the test fails with:

=== NAME  TestMonitorExecution/inject-error
    vm_test.go:431: got no report

Make the command execute for a second so that output
is always injected.