pkg/ifuzz: fix instruction decoding on x86
Decode() was only checking full opcode byte(s), whereas certain
instructions are encoded in a way that some bits of the opcode are
stored in the ModR/M byte.

In particular, e.g. there is a variation of MUL encoded as:
  F7 /4
(which means the opcode byte is F7, and MODRM.reg is 4), and a variation
of TEST encoded as:
  F7 /0
(opcode byte is also F7, and MODRM.reg is 0), which were previously
indistinguishable (the decoder would incorrectly treat the MUL instruction
as a TEST instruction if there were at least four extra bytes following it).

Make sure to calculate and check the MODRM.reg value if insn.Reg is set to a
non-negative value.