syz-manager: fix up checkDone checks
It's possible that a VM crashes after checking was done
but before the corpus is loaded, check checkDone after
the serv.mu to avoid this.

Also double-check checkDone in ExchangeInfo under serv.mu.
It's possible that the check is being finished, but checkDone
is not set yet. It looks like nothing bad happened by accident,
but this whole race wasn't intentional.