pkg/repro: avoid hitting the prog.MaxCalls limit
When we combine the progs found during prog bisection, there's a chance
that we may exceed the prog.MaxCalls limit. In that case, we get a
SYZFATAL error and proceed as if it were the target crash. That's
absolutely wrong.

Let's first minimize each single program before concatenating them, that
should work for almost all cases.