syz-manager: check GitRevisionBase length
In the summary page of the http server, the GitRevisionBase is truncated
to 8 characters and displayed in the UI. However, when building
syzkaller through other build systems, this value is not properly set
and remains "unknown".

Unfortunately, "unknown" has only 7 letters, which causes the code that
truncates to 8 letters to panic, and the UI page can't be shown.

To fix this without doing any more invasive changes to the logic behind
how GitRevisionBase is set, we will simply check if it has more than 8
characters, and truncate it only if it does. Otherwise we will display
it in full.