executor: fix extraction of number of KCOV comparisons from coverage data
KCOV stores the number of KCOV comparisons in a coverage buffer always
as a 64-bit integer at offset 0 of the coverage buffer. Don't use
the field size of the coverage object which is initialized
in cover_collect() and size of which depends on kernel bitness because
this field is intended only for KCOV PC coverage and not for
KCOV comparisons.

Signed-off-by: Alexander Egorenkov <eaibmz@gmail.com>