Revert "syz-manager: support stdin as port forwarding result"
This reverts commit 98cac403c442a32da968c5ef3c0d96b974fcf9bb.

The gvisor "stdin" address still crashes in executor
Connection::Connect on atoi(ports) with ports == NULL.

The gvisor "stdin" address is not tested, so it's better to make it less
special rather than add more special cases in manager, executor,
and now also in Connection to handle it.
It still may crash in future after some changes.