pkg/vminfo: make it possible to force coverage support
We need this in pkg/runtest since not all TestOS targets natively
support the coverage instrumentation.

We used to achieve this by starting the RPC server with Coverage=false
and then updating it to Coverage=true to suppress fallback the fallback
signal, but it's better to avoid such runtime config changes.