syz-manager: query hub only after finishing the previous batch
It does happen that we see a long tail of "candidate triage jobs"
during a big influx of syz-hub programs. This is bad because in 10
minutes we'll query another batch, which will further stretch the triage
process.

By delaying the process more and more we're offloading the start of bug
reproduction, so let's control the hub sync process more carefully -
only perform the next query after the previous batch has completed.