executor: deduplicate signal per-call
This kind of deduplication is confusing for the fuzzer, which expects to
control the process itself (by MaxSignal and by specifying the calls for
which full signal must be returned).

There's also a chance that it may contribute to the difficulties during
program triage and minimization.

Let's err on the safe side and deduplicate signal only per-call.