prog: don't minimize int/resource for corpus
It makes little sense to minimize int's for corpus.
Also replacing resource with a default value does not make sense as well.
For corpus we are only interesting in reducing total number of args
that will be considered for mutation.

Add CrashSnapshot mode, mainly to keep the minimization code "alive" for now.